Added js files to html
