+++ ps3/index.html
@@ -48,5 +48,6 @@
     </div>
 
     <script type="module" src="/public/main.js"></script>
+    <script type="module" src="/public/html_modification.js"></script>
   </body>
 </html>
