+++ ps3/public/html_modification.js
@@ -1,0 +1,5 @@
+// let key = document.getElementById("key");
+// key.addEventListener('keyup', function(evt){
+//     var n = parseInt(this.value.replace(/\D/g,''), "-");
+//     key.value = n;
+// }, false);
